Elias-Fano now is usize

+++ src/graph/bvgraph/bvgraph_random_access.rs
@@ -7,7 +7,7 @@ use crate::utils::CircularBufferVec;
 
 /// BVGraph is an highly compressed graph format that can be traversed
 /// sequentially or randomly without having to decode the whole graph.
-pub struct BVGraph<CRB: BVGraphCodesReaderBuilder, OFF: IndexedDict<Value = u64>> {
+pub struct BVGraph<CRB: BVGraphCodesReaderBuilder, OFF: IndexedDict<Value = usize>> {
     /// Backend that can create objects that allows us to read the bitstream of
     /// the graph to decode the edges.
     codes_reader_builder: CRB,
@@ -27,7 +27,7 @@ pub struct BVGraph<CRB: BVGraphCodesReaderBuilder, OFF: IndexedDict<Value = u64>
 impl<CRB, OFF> BVGraph<CRB, OFF>
 where
     CRB: BVGraphCodesReaderBuilder,
-    OFF: IndexedDict<Value = u64>,
+    OFF: IndexedDict<Value = usize>,
 {
     /// Create a new BVGraph from the given parameters.
     ///
@@ -84,7 +84,7 @@ where
     pub fn map_offsets<OFF2, F>(self, map_func: F) -> BVGraph<CRB, OFF2>
     where
         F: FnOnce(MemCase<OFF>) -> MemCase<OFF2>,
-        OFF2: IndexedDict<Value = u64>,
+        OFF2: IndexedDict<Value = usize>,
     {
         BVGraph {
             codes_reader_builder: self.codes_reader_builder,
@@ -106,7 +106,7 @@ where
 impl<CRB, OFF> SequentialGraph for BVGraph<CRB, OFF>
 where
     CRB: BVGraphCodesReaderBuilder,
-    OFF: IndexedDict<Value = u64>,
+    OFF: IndexedDict<Value = usize>,
 {
     type NodesIter<'b> = WebgraphSequentialIter<CRB::Reader<'b>>
         where Self: 'b, CRB: 'b,
@@ -164,7 +164,7 @@ where
 impl<CRB, OFF> RandomAccessGraph for BVGraph<CRB, OFF>
 where
     CRB: BVGraphCodesReaderBuilder,
-    OFF: IndexedDict<Value = u64>,
+    OFF: IndexedDict<Value = usize>,
 {
     type RandomSuccessorIter<'b> = RandomSuccessorIter<CRB::Reader<'b>>
         where Self: 'b,
@@ -403,7 +403,7 @@ impl<CR: BVGraphCodesReader> Iterator for RandomSuccessorIter<CR> {
 impl<'a, CRB, OFF> IntoIterator for &'a BVGraph<CRB, OFF>
 where
     CRB: BVGraphCodesReaderBuilder,
-    OFF: IndexedDict<Value = u64>,
+    OFF: IndexedDict<Value = usize>,
 {
     type IntoIter = WebgraphSequentialIter<CRB::Reader<'a>>;
     type Item = <WebgraphSequentialIter<CRB::Reader<'a>> as Iterator>::Item;
